Move methods out of the GameTeam class
